refactors new tests
